doc(controller) documents POST /v2/auth/register
This varies very slightly from the actual behavior of the Python
controller, in so far as the `POST` payload types are strict and
the `email` field is considered required.

The actual behavior in the Python app is that email is technically
required (omiting it produces a server error), but it is only
validated if it is included.